Create TestEffects.py
Create a simple program to be used to test individual (or groups of effects running together).

To test your effects, implement as a class modeled after the WipeFillEffect.    Then insert your effects so they are in the list returned by the Effects Chooser (in the function get_chosen_effects).   If you only have one effect to test, it should work to put it as the only element in the list.